pyproject.toml
[project]
name = "neural-inducing-kernel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kornia>=0.7.3",
    "lightning>=2.4.0",
    "pillow>=11.0.0",
    "tensorboard>=2.18.0",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
]
